requirements.txt
fastapi
uvicorn
groq
requests
pillow
pytesseract
chromadb
sentence-transformers
